Renamed the DirectedWeightedGraph to DirectedWeightedMultigraph, Graph to Multigraph, and DirectedWeightedGraphTest to DirectedWeightedMultigraphTest to better reflect the data structure.
